Cleaning up; moving to new Candidate objects
