+++ snorkel/matcher_utils.py
@@ -1,0 +1,24 @@
+import numpy as np
+import re
+
+def tag_seq(words, seq, tag):
+  """Sub in a tag for a subsequence of a list"""
+  words_out = words[:seq[0]] + ['{{%s}}' % tag]
+  words_out += words[seq[-1] + 1:] if seq[-1] < len(words) - 1 else []
+  return words_out
+
+def tag_seqs(words, seqs, tags):
+  """
+  Given a list of words, a *list* of lists of indexes, and the corresponding tags
+  This function substitutes the tags for the words coresponding to the index lists,
+  taking care of shifting indexes appropriately after multi-word substitutions
+  NOTE: this assumes non-overlapping seqs!
+  """
+  words_out = words
+  dj = 0
+  for i in np.argsort(seqs, axis=0):
+    i = int(i[0]) if hasattr(i, '__iter__') else int(i)
+    words_out = tag_seq(words_out, map(lambda j : j - dj, seqs[i]), tags[i])
+    dj += len(seqs[i]) - 1
+  return words_out
+
